backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.11.0
pydantic-settings==2.1.0
google-generativeai>=0.8.0
pinecone-client>=3,<4
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-pinecone==0.0.1
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1 